+++ src/api/endpoints/execute/get-execute-sell/v3.ts
@@ -154,7 +154,7 @@ export const getExecuteSellV3Options: RouteOptions = {
           contract,
           tokenId,
           quantity: 1,
-          source: sourceId ? sources.get(sourceId) : null,
+          source: sourceId ? sources.get(sourceId).domain : null,
           // TODO: Add support for multiple currencies
           currency: Sdk.Common.Addresses.Weth[config.chainId],
           quote: formatEth(bestOrderResult.price),